requirements.txt
flask==3.1.0
gunicorn==23.0.0
streamlit==1.40.0
pypdf==5.1.0
langchain==0.3.0
langchain-google-genai==2.1.12
langchain-core==0.3.0
google-generativeai==0.8.3
python-dotenv==1.0.1